modified tmux and zshrc

+++ zsh/.zshrc
@@ -198,9 +198,28 @@ _fzf_compgen_dir() {
 }
 
 # ------------ bat ----------------
-if [[ ! -f ~/.local/bin/bat ]] 
-then
+if [[ ! -f ~/.local/bin/bat && -f /usr/bin/batcat ]]; then
     ln -s /usr/bin/batcat ~/.local/bin/bat
 fi
 export BAT_THEME=tokyonight_night
+alias cat=bat
+
+# ---------- fzf with eza and bat ----------
+export FZF_CTRL_T_OPTS="--preview 'bat -n --color=always --line-range :500 {}'"
+export FZF_ALT_C_OPTS="--preview 'eza --tree --color=always {} | head -200'"
+
+# Advanced customization of fzf options via _fzf_comprun function
+# - The first argument to the function is the name of the command.
+# - You should make sure to pass the rest of the arguments to fzf.
+_fzf_comprun() {
+  local command=$1
+  shift
+
+  case "$command" in
+    cd)           fzf --preview 'eza --tree --color=always {} | head -200' "$@" ;;
+    export|unset) fzf --preview "eval 'echo $'{}"         "$@" ;;
+    ssh)          fzf --preview 'dig {}'                   "$@" ;;
+    *)            fzf --preview "bat -n --color=always --line-range :500 {}" "$@" ;;
+  esac
+}
